project section edit, update, delete

+++ assets/js/task.js
@@ -1,5 +1,6 @@
 const task_table = document.getElementById(`task_table`);
 const hiding_over_task = document.getElementById(`hiding_over_task`);
+const resource_pool = document.getElementById(`resource_pool`);
 
 // Top Counters
 const total_task_count = document.getElementById(`total_task_count`);
@@ -29,7 +30,9 @@ let current_editing_task = 1;
 // Edit Task
 const edit_task = document.getElementById(`edit_task`);
 
-//////////// MAKE TASK TABLE READY ///////
+const delete_overlay_task = document.getElementById(`delete_overlay_task`);
+
+// MAKE TASK TABLE READY
 const make_task_table_ready = () => {
   let total_task = 0,
     pending_task = 0,
@@ -46,7 +49,7 @@ const make_task_table_ready = () => {
                 <th style="width: 8%">Action</th>
             </tr>`;
 
-  fetch(`http://localhost:3005/api/get_task`)
+  fetch(`${localhost}/api/get_task`)
     .then((response) => response.json())
     .then((data) => {
       total_task = data.length;
@@ -95,9 +98,10 @@ const make_task_table_ready = () => {
     });
 
   // Render Employee list
-  fetch(`http://localhost:3005/api/get_employee`)
+  fetch(`${localhost}/api/get_employee`)
     .then((response) => response.json())
     .then((data) => {
+      let temp = ``;
       if (data.length > 0) {
         new_task_user.innerHTML = "";
         data.forEach((employee) => {
@@ -105,6 +109,31 @@ const make_task_table_ready = () => {
           new_option.text = employee.name;
           new_option.value = employee.id;
           new_task_user.append(new_option);
+
+          // Resource Pool
+          fetch(`${localhost}/api/task_with_employeeid`, {
+            method: "POST",
+            body: JSON.stringify({
+              id: employee.id,
+            }),
+            headers: {
+              "Content-type": "application/json; charset=UTF-8",
+            },
+          })
+            .then((response) => response.json())
+            .then((data) => {
+              temp += `<div id="${
+                employee.id
+              }" class="border-bottom resource-txt py-2 rounded text-center my-flex-row">
+                    <span class="pt-2 ${
+                      data.length == 0 ? "make_green fa-fade" : ""
+                    }">${employee.name} </span>
+                    <span class="pt-2" >
+                    <span class="badge badge-count">${data.length} Task</span>
+                    </span>
+                  </div>`;
+              resource_pool.innerHTML = temp;
+            });
         });
       } else {
         let new_option = document.createElement("option");
@@ -116,7 +145,7 @@ const make_task_table_ready = () => {
     });
 
   // Render Project List
-  fetch(`http://localhost:3005/api/get_projects`)
+  fetch(`${localhost}/api/get_projects`)
     .then((response) => response.json())
     .then((data) => {
       if (data.length > 0) {
@@ -137,15 +166,46 @@ const make_task_table_ready = () => {
     });
 };
 
-make_task_table_ready();
+// GET CONFIRMATION FOR DELETION
+const get_delete_confirmation = () => {
+  return new Promise((resolve) => {
+    // Get the confirmation and cancel buttons
+    const confirmationButton = document.getElementById("proceed_task_delete");
+    const cancelButton = document.getElementById("cancel_task_delete");
 
-/////////// DELETE TASK //////////
-const delete_task = (event) => {
+    // Function to handle confirmation
+    const handleConfirmation = () => {
+      resolve(true); // Resolve with true when confirmation button is clicked
+    };
+
+    // Function to handle cancellation
+    const handleCancellation = () => {
+      resolve(false); // Resolve with false when cancel button is clicked
+    };
+
+    // Attach click event listeners to buttons
+    confirmationButton.addEventListener("click", handleConfirmation);
+    cancelButton.addEventListener("click", handleCancellation);
+  });
+};
+
+// DELETE TASK
+const delete_task = async (event) => {
+  event.stopPropagation();
+  delete_overlay_task.style.display = "block";
+
+  let confirmation = await get_delete_confirmation();
+  delete_overlay_task.style.display = "none"; // Hide SweetAlert
+
+  if (!confirmation) {
+    return;
+  }
+
   hiding_over_task.style.display = "block";
   let parent = event.target;
   while (parent.id == ``) parent = parent.parentNode;
 
-  fetch("http://localhost:3005/api/delete_task", {
+  fetch(`${localhost}/api/delete_task`, {
     method: "DELETE",
     body: JSON.stringify({
       id: parent.id,
@@ -190,11 +250,55 @@ const delete_task = (event) => {
     });
 };
 
-///////// ADD NEW TASK //////////////
+// ADD NEW TASK
 const save_new_task = () => {
+  const formElements = [
+    new_task_name,
+    new_task_description,
+    new_task_user,
+    new_task_priority,
+    due_date_task,
+    new_task_project,
+  ];
+
+  let validator = false;
+
+  formElements.forEach((element) => {
+    if (element.value === "") {
+      element.style.border = "1px solid #f58389";
+      validator = true;
+    } else {
+      element.style.border = "1px solid #6ddc70";
+    }
+  });
+
+  if (validator) {
+    // Notification
+    let content = {
+      message: "Please fill out all fields correctly to add a new Task.",
+      title: "Incomplete Task Details",
+      icon: "fas fa-tasks",
+      url: "task.html",
+      target: "_blank",
+    };
+
+    $.notify(content, {
+      type: "warning",
+      placement: {
+        from: "top",
+        align: "center",
+      },
+      time: 100,
+      delay: 1000,
+    });
+
+    return;
+  }
+
   hide_add_task_form();
-  hiding_over_task.style.display = "none";
-  fetch("http://localhost:3005/api/add_task", {
+  hiding_over_task.style.display = "block";
+
+  fetch(`${localhost}/api/add_task`, {
     method: "POST",
     body: JSON.stringify({
       name: new_task_name.value,
@@ -216,7 +320,8 @@ const save_new_task = () => {
       };
       let alert_type = "";
       if (json.success) {
-        content.message = "A new Task has been added to the database.";
+        content.message =
+          "A new task has been successfully added to the database.";
         content.title = "Task Added";
         content.icon = "fas fa-tasks";
         alert_type = "success";
@@ -243,12 +348,55 @@ const save_new_task = () => {
     });
 };
 
-/////////// EDIT TASK //////////////
+// EDIT TASK
 const update_task = () => {
+  const formElements = [
+    edit_task_name,
+    edit_task_description,
+    edit_task_user,
+    edit_task_priority,
+    edit_date_task,
+    edit_task_project,
+  ];
+
+  let validator = false;
+
+  formElements.forEach((element) => {
+    if (element.value === "") {
+      element.style.border = "1px solid #f58389";
+      validator = true;
+    } else {
+      element.style.border = "1px solid #6ddc70";
+    }
+  });
+
+  if (validator) {
+    // Notification
+    let content = {
+      message: "Please fill out all fields correctly to Edit Task.",
+      title: "Incomplete Task Details",
+      icon: "fas fa-tasks",
+      url: "task.html",
+      target: "_blank",
+    };
+
+    $.notify(content, {
+      type: "warning",
+      placement: {
+        from: "top",
+        align: "center",
+      },
+      time: 100,
+      delay: 1000,
+    });
+
+    return;
+  }
+
   hide_edit_task_form();
   hiding_over_task.style.display = "block";
 
-  fetch("http://localhost:3005/api/edit_task_with_id", {
+  fetch(`${localhost}/api/edit_task_with_id`, {
     method: "PUT",
     body: JSON.stringify({
       id: current_editing_task,
@@ -298,46 +446,30 @@ const update_task = () => {
     });
 };
 
-///////// HIDE ADD NEW TASK FORM ////////////
+// HIDE ADD NEW TASK FORM
 const hide_add_task_form = () => {
   add_new_task.style.display = "none";
 };
 
-///////// VIEW ADD NEW TASK FORM ////////////
+// VIEW ADD NEW TASK FORM
 const view_add_task_form = () => {
   add_new_task.style.display = "block";
 };
 
-///////// HIDE EDIT TASK FORM ////////////
+// HIDE EDIT TASK FORM
 const hide_edit_task_form = () => {
   edit_task.style.display = "none";
 };
 
-///////// VIEW EDIT TASK FORM ////////////
+// VIEW EDIT TASK FORM
 const view_edit_task_form = (event) => {
+  event.stopPropagation();
   edit_task_user.innerHTML = "";
   let parent = event.target;
   while (parent.id == ``) parent = parent.parentNode;
 
-  // Fetch Task List With ID
-  fetch("http://localhost:3005/api/get_task_with_id", {
-    method: "POST",
-    body: JSON.stringify({
-      id: parent.id,
-    }),
-    headers: {
-      "Content-type": "application/json; charset=UTF-8",
-    },
-  })
-    .then((response) => response.json())
-    .then((data) => {
-      edit_task_name.value = data[0].name;
-      edit_task_description.value = data[0].description;
-      edit_date_task.value = data[0].due_date.slice(0, 10);
-    });
-
   // Render Employee list
-  fetch(`http://localhost:3005/api/get_employee`)
+  fetch(`${localhost}/api/get_employee`)
     .then((response) => response.json())
     .then((data) => {
       if (data.length > 0) {
@@ -357,7 +489,7 @@ const view_edit_task_form = (event) => {
     });
 
   // Render Project List
-  fetch(`http://localhost:3005/api/get_projects`)
+  fetch(`${localhost}/api/get_projects`)
     .then((response) => response.json())
     .then((data) => {
       if (data.length > 0) {
@@ -376,7 +508,28 @@ const view_edit_task_form = (event) => {
       }
     });
 
+  // Fetch Task List With ID
+  fetch(`${localhost}/api/get_task_with_id`, {
+    method: "POST",
+    body: JSON.stringify({
+      id: parent.id,
+    }),
+    headers: {
+      "Content-type": "application/json; charset=UTF-8",
+    },
+  })
+    .then((response) => response.json())
+    .then((data) => {
+      edit_task_name.value = data[0].name;
+      edit_task_description.value = data[0].description;
+      edit_date_task.value = data[0].due_date.slice(0, 10);
+      edit_task_priority.value = data[0].priority;
+      edit_task_user.value = data[0].employee_id;
+      edit_task_project.value = data[0].project_id;
+    });
+
   current_editing_task = parent.id;
-  console.log(current_editing_task);
   edit_task.style.display = "block";
 };
+
+make_task_table_ready();
